Feat: 포토카드 이미지 자르기 기능에서 일부 부분 이미지만 삭제할 수 있는 기능 추가
