+++ client/src/page/admin/photo/crop/cropCard/Crop.tsx
