+++ client/src/page/admin/photo/crop/cropCard/Index.tsx
